// src/app/models/cuidador.ts
export class Cuidador  {
    id: number;
    tipoIdentificacion: string;
    numeroIdentificacion: number;
    primerNombre: string;
    segundoNombre: string;
    primerApellido: string;
    segundoApellido: string;
    parentesco: string;    
    correoElectronico: string;
    indicativoTelefono: number;
    telefonoFijo: number;
    celular: number;
  }